# routes/display/Display.js
// All routes related to club's LOGIN AND REGISTER

const express = require("express");
const User = require("../../schema/user/UserSchema");
const { STATUSCODE, STATUSMESSAGE } = require("../../utils/Status");
const Event = require("../../schema/events/EventSchema");
const router = express.Router();

// Route 1  - Show all avaialble Users in the DB
router.get("/users", async (req, res) => {
  try {
    const allusers = await User.find({});
    res.status(STATUSCODE.OK).json(allusers);
  } catch (error) {
    res.status(500).json({ message: "Internal Server Error" });
  }
});

// * Route 2 - Show all available Clubs in the DB
router.get("/clubs", async (req, res) => {
  try {
    const allClubs = await User.find({ usertype: "club" });
    res.json(allClubs);
  } catch (error) {
    res.status(500).json({ message: "Internal Server Error" });
  }
});

// * Route 3 - Show all available Events in the DB
router.get("/events", async (req, res) => {
  try {
    const allEvents = await Event.find();
    res.status(STATUSCODE.OK).json(allEvents);
  } catch (err) {
    res
      .status(STATUSCODE.INTERNAL_SERVER_ERROR)
      .json(STATUSMESSAGE.INTERNAL_SERVER_ERROR);
  }
});

module.exports = router;